+++ src/Generics/CompositeArithmetics.cu
@@ -18,7 +18,7 @@ template <class T> __global__ void SquaredDistanceFromScalarKernel(T* d_input, T
 template <class T> void d_SquaredDistanceFromVector(T* d_input, T* d_vector, T* d_output, size_t elements, int batch)
 {
 	size_t TpB = 256;
-	size_t totalblocks = min((elements + TpB - 1) / TpB, 32768);
+	size_t totalblocks = min((elements + TpB - 1) / TpB, (size_t)32768);
 	dim3 grid = dim3((uint)totalblocks);
 	SquaredDistanceFromVectorKernel<T> <<<grid, (uint)TpB>>> (d_input, d_vector, d_output, elements, batch);
 }
@@ -28,7 +28,7 @@ template void d_SquaredDistanceFromVector<int>(int* d_input, int* d_vector, int*
 template <class T> void d_SquaredDistanceFromScalar(T* d_input, T* d_output, size_t elements, T scalar)
 {
 	size_t TpB = 256;
-	size_t totalblocks = min((elements + TpB - 1) / TpB, 32768);
+	size_t totalblocks = min((elements + TpB - 1) / TpB, (size_t)32768);
 	dim3 grid = dim3((uint)totalblocks);
 	SquaredDistanceFromScalarKernel<T> <<<grid, (uint)TpB>>> (d_input, d_output, elements, scalar);
 }
@@ -38,7 +38,7 @@ template void d_SquaredDistanceFromScalar<int>(int* d_input, int* d_output, size
 template <class T> void d_SquaredDistanceFromScalar(T* d_input, T* d_scalars, T* d_output, size_t elements, int batch)
 {
 	size_t TpB = 256;
-	size_t totalblocks = min((elements + TpB - 1) / TpB, 32768);
+	size_t totalblocks = min((elements + TpB - 1) / TpB, (size_t)32768);
 	dim3 grid = dim3((uint)totalblocks, batch);
 	SquaredDistanceFromScalarKernel<T> <<<grid, (uint)TpB>>> (d_input, d_scalars, d_output, elements);
 }
